Mona Alkhammash
on a scale from 1 to 5, how comfortable were you with this assignment?
5

What was a win you had with this assignment?
new code ( slides Up and down|) always practicing.

What was a challenge you had with this assignment?
Nothing

Is there anything that you'd like some further information on?
Nothing

Other stuff on your mind? work with anyone? tell us about it!
Thank you everthing